chatbot-api/requirements.txt
fastapi
uvicorn[standard]
qdrant-client[fastembed]
python-dotenv
openai
